rename ElasticsearchOutput to TypedSearchResponse
