+++ tests/shared.ts
@@ -1,9 +1,6 @@
 import type { estypes } from "@elastic/elasticsearch";
-import type {
-	ElasticsearchOutput,
-	TypedClient,
-	TypedSearchRequest,
-} from "../src/index";
+import type { TypedClient, TypedSearchRequest } from "../src/index";
+import type { TypedSearchResponse } from "../src/override/search-response";
 
 export type SearchRequest = estypes.SearchRequest;
 
@@ -75,7 +72,7 @@ export type TestAggregationOutput<
 	Index extends keyof CustomIndexes,
 	Aggregation extends TypedSearchRequest<CustomIndexes>["aggregations"],
 > = {
-	aggregations: ElasticsearchOutput<
+	aggregations: TypedSearchResponse<
 		{
 			index: Index;
 			rest_total_hits_as_int: true;
